Convert CoglTextureGLVertex to be GLfloat instead of GLfixed
This causes vertices from cogl_texture_rectangle and
cogl_texture_polygon to be sent as floating point values rather than
fixed point which is also what happens in current master.

The PowerVR simulator appears to have a bug that if you send data as
GL_FIXED in combination with glDrawElements then it will fail to
convert the last vertex to float before sending to GL and it will be
garbage. The changes in the merged multiple-texture-rectangle branch
cause this situation.