Rework the viewport resizing
The patch in commit a93c0a6a5 was overly complex. The correct
and simplest approach is to call XResizeWindow() and not set the
SYNC_MATRICES private flag -- thus avoiding a maybe_setup_viewport()
call.

The SYNC_MATRICES flag is set only after the stage Drawable received
a ConfigureNotify event.