Add gen-changelog.pl
The gen-changelog.pl is a simple script for generating the ChangeLog
file out of the git log.

It should be called prior to `make dist`, like:

        $ perl build/gen-changelog.pl > ChangeLog

This will generate the ChangeLog since the first commit object. If
you want to generate the ChangeLog starting from a revision or a
particular tag, use something like:

        $ perl build/gen-changelog.pl 0.8.4 > ChangeLog