Update Clutter to use G-I 0.6.4
Following bug #1762, the syntax of g-ir-scanner was changed in
gobject-introspection, so Clutter does not build anymore with 0.6.4.

See the bugzilla bug:

  http://bugzilla.gnome.org/show_bug.cgi?id=591669

GObject-Introspection now uses a different mechanism to extract the
SONAME when building the gir file and it needs the libtool archive as
option.

Since there has been a little bit of back and forth on this patch, this
commit squashes the following commits on master:

  366e0f9cd7773433ba14cc531119bca3b9879a1e Götz Waschk
  68f8a98cfbf70e2b7bfcda7831907633eb08ca50 Götz Waschk
  4c1248c5e9b934b23bdabbfc3d2dabdb8bab0656 Owen W. Taylor