Clarify the ID pool warnings
Instead of having an assertion failure with a message of dubious
usefulness, we should probably use a more verbose warning explaining
what is the problem and what might be the cause.

We should also warn if the glReadPixels happens to generate an ID that
is within the range of the ID array but no longer points to an
existing actor.

Cherry picked from master:

  ad83caaed16a36aea0fafd4ed87cc283c30c413b Emmanuele Bassi
  a4ff4d37a40b60ce7f7029c6fe12181c0e1e89e5 Neil Roberts
  13ac1fe75bec418f4e37e04540e6d16af55413f5 Neil Roberts