actor: Show actor name or type in the state checks warnings
When showing a warning in the state checks we perform to verify that
the invariants are maintained when showing, mapping and realizing, we
should also print out the name of the actor failing the checks. If the
actor has no name, the GType name should be used as a fallback.
(cherry picked from commit 7de47e74bda68c5d4dde82f2cbba6782fcf69d27)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>