actor: Do not use G_UNLIKELY in paint()
We decide whether the paint() should be a real paint or a paint in pick
mode depending on the global pick_mode value. Using G_UNLIKELY() on an
operation that most likely is going to be executed once every frame is
going to blow a lot of cache lines and frak with the CPU branch
prediction. Not good.
(cherry picked from commit 93a5b78e5ae3ab07c11f92fc6d7525dc3569dcb4)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>