x11: Use g_signal_override_class_closure()
The g_signal_override_class_handler() function was added in GLib 2.18,
which means we should bump our requirement for GLib. Luckily, the
override_class_handler() function is just a wrapper around the older
g_signal_override_class_closure(), so we can use that and avoid bumping
the pre-requisites in the middle of a stable cycle.