cogl-texture: Don't attempt to use GL to convert formats under GLES
In commit abe91784c4b I changed cogl-texture so that it would use the
OpenGL mechanism to specify a different internal texture format from
the image format so that it can do the conversion instead of
Cogl. However under GLES the internal format and the image format must
always be the same and it only supports a limited set of formats. This
patch changes _cogl_texture_prepare_for_upload so that it does the
conversion using the cogl bitmap code when compiling for GLES.

http://bugzilla.openedhand.com/show_bug.cgi?id=2059
(cherry picked from commit 405d529f449bb3cfa7f156510a98e70e38c9e1b5)