test-conformance: Use -export-dynamic instead of -rdynamic
The -rdynamic linker option is specific to ELF so it was breaking
builds on systems with other object formats such as Windows and
Solaris. This patch replaces that option with -export-dynamic which is
a portable libtool option which should do the right thing on each
platform.

http://bugzilla.clutter-project.org/show_bug.cgi?id=1930
(cherry picked from commit fbeab378289a0e9f9470aa9f532003cf66529970)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>