clutter-offscreen-effect: Don't recreate the material when FBO changes
Previously whenever the size of the FBO changes it would create a new
material and attach the texture to it. This is not good for Cogl
because it throws away any cached state for the material. In
test-rotate the size of the FBO changes constantly so it effectively
uses a new material every paint. For shader effects this also ends up
relinking the shader every paint because the linked programs are part
of the material state.
(cherry picked from commit 6ddb8397110f6e6998e0262e5d3dd604cd93d670)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>