ClutterShaderEffect: Don't throw away the program for a new actor
Previously when the shader effect is used with a new actor it would
end up throwing away the old program. I don't think this is neccessary
and it means if you use an effect to temporarily bind to an actor then
it will recompile the shader whenever it is applied.

http://bugzilla.clutter-project.org/show_bug.cgi?id=2454
(cherry picked from commit 08087c965bf312513f98a53bf3220e92c372d131)

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>