Add GObject accessor macros
Adding accessors for simple properties in GObject is a huge chore.

Instead of a full blown IDL, we could use a bunch of pre-processor
macros, in the same spirit of cutting down boilerplate code as
G_DEFINE_TYPE and friends.

Without further ado, here's to you:

  ‣ G_DEFINE_PROPERTY()
  - a macro for declaring accessors in header files

  ‣ G_DEFINE_PROPERTY_SYNTH()
  - a macro for defining accessors in source files

  ‣ G_DEFINE_PROPERTY_SYNTH_SET() and
    G_DEFINE_PROPERTY_SYNTH_SET_WITH_CODE()
  - two macros for defining the body of a setter function, with
    and without custom code

  ‣ G_DEFINE_PROPERTY_SYNTH_GET() and
    G_DEFINE_PROPERTY_SYNTH_GET_WITH_CODE()
  - two macros for defining the body of a getter function, with
    and without custom code

For the time being, these macros are private, and should only be
used internally — at least until they get moved to GObject.