Snapshot/WIP
The AnimationContext is a meta-class that can be used by Animatable
implementations to keep track of whether they are currently being
used in an animation; the names and intervals of the properties
currently being animated; the duration of the animation and its
easing mode.

At the beginning of an animation, the Animatable object is asked to
create an appropriate AnimationContext instance; the animation code
should then describe the animation itself; and finally it should
push the animation context on top of a stack (maintained by the
Animatable object itself).

The Animatable implementation is now able to verify, when changing
its state, if there is an animation in progress, and can defer any
state change to the end of the frame; this way, work like signal
emission and property notification can be coalesced at the end of
the frame.