Add clutter_shutdown()
Most of the EGL platforms have a requirement to dispose the Clutter
backend, in order to restore the graphics stack to an uninitialized
state.

Up until Clutter 1.8 we had a fairly hacky atexit() handler that called
g_object_unref() on the backend singleton — but apart from debugging
purposes and inside a controlled environment, atexit() handlers should
not be used inside library code.

Since we guarantee that the Clutter main loop can be stopped and
restarted at any point without any loss of state, if we want to allow
developers of specific platforms to clean up after the end of the
process, then we need a new symbol — a "deinitialization" function
bookending clutter_init().

The new symbol is called clutter_shutdown().

It's important to note that this function is not going to be useful for
debugging or for memory profiling, e.g. for cleaning up Valgrind output;
Clutter, and all the libraries that it depends on, does a set of one-off
allocations that are better isolated by using a proper suppression file.

https://bugzilla.gnome.org/show_bug.cgi?id=664051